fix: resolve 4 critical bugs (#17, #18, #19, #21) (#30)
- #18: Delete unauthenticated create-market-metadata edge function
  (service_role upsert with no auth, unused by frontend)

- #19: Fix ClaimModal to pass correct 4 params (u128) to claim_winnings
  Also fix OracleResolveModal pool params from u64 to u128

- #21: Remove frontend upsertUserPosition call from BetModal + dead code
  (both frontend + indexer were incrementing = double-counted positions)

- #17: Replace read-then-write incrementPoolTotal with atomic Postgres RPC
  (concurrent bets could permanently lose pool totals via last-write-wins)

- Clean up stale comments and update README

Migration 0004_atomic_pool_increment.sql (already applied to production DB).

+++ frontend/src/components/OracleResolveModal.tsx
@@ -54,8 +54,8 @@ export function ResolveModal({
           inputs: [
             marketId,
             `${outcome}`,
-            `${yesPool}u64`,
-            `${noPool}u64`,
+            `${yesPool}u128`,
+            `${noPool}u128`,
             `${sumBlindingYes}scalar`,
             `${sumBlindingNo}scalar`,
           ],
